Change return type from int to float for ABF op

+++ famapy/core/operations/average_branching_factor.py
@@ -10,5 +10,5 @@ class AverageBranchingFactor(Operation):
         pass
 
     @abstractmethod
-    def get_average_branching_factor(self) -> int:
+    def get_average_branching_factor(self) -> float:
         pass
